Pull http services into separate directory with barrel file
