+++ frontend/src/app.tsx
@@ -3,9 +3,11 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { AppRouter } from './router';
 import { AuthDataServiceContextProvider } from './services/auth-data-service';
-import { AuthHttpServiceContextProvider } from './services/auth.http-service';
-import { HttpServiceContextProvider } from './services/http-service';
-import { ServiceDocsHttpServiceContextProvider } from './services/service-docs.http-service';
+import {
+  AuthHttpServiceContextProvider,
+  HttpServiceContextProvider,
+  ServiceDocsHttpServiceContextProvider,
+} from './services/http';
 
 export const App: React.FC = () => {
   return (
