+++ frontend/src/components/login-page/login-page.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { APP_ROUTES } from '../../routes';
-import { useAuthHttpServiceContext } from '../../services/auth.http-service';
+import { useAuthHttpServiceContext } from '../../services/http';
 
 enum ViewState {
   Default,
